refactor: class TIL 정리
