+++ TIL/daily/2025-03-14.md
@@ -1,36 +1,50 @@
+# 레퍼런스 타입과 참조
+
+- 가상세계 = 컴퓨터의 메모리 영역
+- 인스턴스 = `heap` 영역 안에 확보한 메모리
+- Dart 는 모든 타입이 레퍼런스 타입
+
+---
+
 ```dart
 min(작은 숫자,큰 숫자);
 ```
-      
-      
+
+
+
 **주석 없어도 코드로 알아볼 수 있는 코드가 미덕이다.**
-      
-     
-| 가상세계 | 컴퓨터 메모리 영역 |
-| --- | --- |
-| 인스턴스 | heap 영역 안에 확보된 메모리 |
-|   Dart 에서는 모든 타입이 레퍼런스 타입이다 |
-      
-       
+
+
+
+---
+
+인스턴스화
+
+```bash
+Hero hero = Hero('홍길동', 50);
+```
+
+---
+
 | CODE | 함수, 제어문, 상수  | 컴파일 시 크기가 정함 |
 | --- | --- | --- |
 | DATA | 초기화된 전역변수 | 컴파일 시 크기가 정함 |
 | BSS (Blocked Stated Symbol) | 초기화가 안된 전역변수 | 컴파일 시 크기가 정함 |
 | HEAP | 동적할당 malloc() | 실행 중 메모리 할당 |
 | STACK | 지역변수 | 실행 중 메모리 할당 |
-      
-      
+
+---
+
 ### Heap
-     
-     
+
+
+
 - 내가 사용하고 있는 메모리 양
-       
 - 안드로이드 스튜디오 오른쪽 하단 아래 사용량 확인 가능
-         
 - 메모리 주소는 Dart에서 확인 할 수 없다
-     
-     
-     
+
+---
+
 ```dart
 void main() {
   final hero1 = Hero(name: '홍길동', hp: 100);
@@ -50,23 +64,80 @@ class Hero {
 
 console : 200
 ```
-     
-     
+
+---
+
+### Dart 에서는 모든 타입이 레퍼런스 타입
+
+- `int` 형이나 `double` 형 같은 기본형 (primitive type) 뿐만 아니라 `String` 도 Hero 와 같은 **레퍼런스 타입** 이다.
+
+---
+
+### 생성자 (constructor)
+
+- 인스턴스를 생성하는 방법을 제공
+
+```bash
+final sword = Sword(name: '불꽃의 검', damage: 5);
+```
+
+### named parameter
+
+- 생성자에 `{ }` 를 사용하면 `name parameter`
+- 데이터 타입이 `Null` 을 허용하지 않으며 `required` 를 붙여야 함
+
+### 필수 파라미터와 name parameter (optional)
+
+- 필수 `parameter` 와 `named parameter` 을 동시에 사용할 경우 필수 `parameter` 가 앞에 와야함 마지막 `parameter` 뒤에 `,` 을 찍을 수 없음
+
+```bash
+class Hero {
+  String name;
+  int hp;
+  Sword? sword;
+
+  Hero(this.name, this.hp, {this.sword});
+ }
+```
+
+- 생성 방법
+
+```bash
+Hero hero1 = Hero('전사', 50);
+Hero hero2 = Hero('검사', 100, sword: sword);
+```
+
+- 기본 값 지정
+
+```bash
+class Hero {
+  String name;
+  int hp;
+  Sword? sword;
+
+  Hero(this.name, {this.hp = 100, this.sword});
+ }
+```
+
+---
+
 ### Dart 언어 객체 및 인스턴스 정리
-      
-     
+
+
+
 - 객체 : **현실에서의 실체**를 지칭합니다.
-     
 - 인스턴스 : 가상 환경에서 **class의 구체적인 실체**를 지칭합니다.
-     
+
 ```dart
 void main() {
   final hero1 = Hero(name: '홍길동', hp: 100); // 인스턴스 Hero
 }
 ```
-      
+
+
+
 - 인스턴스 변수 or 맴버 변수
-     
+
 ```dart
 class Hero {
   String name; // 인스턴스 변수 or 맴버 변수 : 클래스의 속성
@@ -75,23 +146,29 @@ class Hero {
   Hero({required this.name, required this.hp});
 }
 ```
-     
+
+
+
 - 클래스의 인스턴스
-     
+
 ```dart
 void main() {
   final hero1 = Hero(name: '홍길동', hp: 100); // Hero class의 instance hero1
   final hero2 = hero1; // Hero class의 instance hero2
 }
 ```
-     
+
+
+
+---
+
 **모든 클래스는 반드시 하나의 생성자를 가진다.**
-     
+
 ```dart
 void main() {
   Hero();
 }
-     
+
 class Hero {
 
 }
@@ -99,9 +176,11 @@ class Hero {
 실행 O 
 비어있는 클래스, 객체 생성 가능
 ```
-     
+
+---
+
 **class 안에 static 을 사용해라**
-     
+
 ```dart
 void main() {
   Hero hero = Hero('홍길동', 30);
@@ -122,33 +201,36 @@ class Hero {
 
 console : 100
 ```
-     
-      
+
+
+
 `class` 와 `static` 은 **물리적으로 서로 다른 영역**에 있다.
-      
+
 `static` 은 `class` 내에서 선언해야하며,
-     
+
 `static` 을 붙이는 순간 class 내에
-     
+
 `static 인스턴스 변수`, `static method` 들은 `class` 것이 아니다.
-     
+
 ```dart
 void main () {
   static String name;
 }
 ```
-     
+
 왜? `main 함수` 에서 못쓰는걸까?
-    
-<img width="464" alt="Image" src="https://github.com/user-attachments/assets/34f40dff-c04a-4ec8-9464-d633b3dd376f" />     
-     
+
+---
+
 https://dart.dev/language/classes#static-variables 공식문서
-     
+
 - You can use static methods as compile-time constants. For example, you can pass a static method as a parameter to a constant constructor.
 - 정적 메서드를 컴파일 타임 상수로 사용할 수 있습니다. 예를 들어 정적 메서드를 상수 생성자에 매개변수로 전달할 수 있습니다.
-      
+
 → 공식 문서가 그렇답니다…
-     
+
+---
+
 **Top level 에서 전역 변수(static) 사용은 자제해라**
 
 ```dart
@@ -161,3 +243,29 @@ class Hero {
   Hero(this.name, this.hp);
 }
 ```
+
+---
+
+### 레퍼런스 타입의 참조
+
+- 레퍼런스 타입 변수의 안에는 “인스턴스 정보가 담겨있는 메모리 번지” 가 들어 있다.
+- 어떤 레퍼런스 타입 변수를 다른 변수에 대입하면, 주소만 복사 된다.
+- 레퍼런스 타입은 필드나 메서드의 인수, 리턴 값의 형으로서도 이용된다.
+
+### 생성자
+
+- “클래스명과 동일 명칭으로, 리턴 값의 타입이 명시되어 있지 않은 메서드"는 생성자로 사용된다.
+- 생성자는 new 에 의한 인스턴스화의 직후에 자동적으로 실행 된다. Dart 에서 new는 생략 가능
+- 인수를 가지는 생성자를 정의하면, new 를 할 때에 인수를 지정하여 생성자를 실행할 수 있다
+- 생성자는 오버로드에 의한 복수 정의가 된다.
+- 클래스에 생성자 정의가 1개도 없는 경우에 한해, 컴파일러가 “인수없음, 처리내용없음"의 기본 생성자를 자동정의 해 준다.
+
+### 정적 멤버
+
+- static 키워드가 붙어 있는 정적 멤버(필드 또는 메소드) 는
+
+    1. 각 인스턴스가 아닌, 클래스에 실체가 준비된다.
+
+    2. 인스턴스를 1개도 생성하지 않아도 이용 가능
+
+- 정적 메서.드는 그 내부에 정적이지 않은 메소드나 필드를 이용하는 것이 불가능하다
